Working test! Can reach into the SHADOW DOM.
