+++ .gitignore
@@ -193,7 +193,7 @@ docs/.docusaurus
 packages/perspective-esbuild-plugin/lib
 docs/static/blocks
 .vscode/c_cpp_properties.json
-/test-results/
-/playwright-report/
-/playwright/.cache/
+test-results/
+playwright-report/
+playwright/.cache/
