+++ packages/perspective-viewer-d3fc/package.json
@@ -41,9 +41,7 @@
         "test:unit": "jest --rootDir=. --roots=test/js/unit --config=../../tools/perspective-test/jest.config.js --color --runInBand",
         "test": "npm-run-all test:build test:run",
         "clean": "rimraf dist",
-        "clean:screenshots": "rimraf \"test/screenshots/**/*.@(failed|diff).png\"",
-        "ptest": "playwright test",
-        "testserver": "npm run ../../scripts/start_testserver.js"
+        "clean:screenshots": "rimraf \"test/screenshots/**/*.@(failed|diff).png\""
     },
     "publishConfig": {
         "access": "public"
