+++ packages/perspective-viewer-d3fc/test/integration/bar.spec.js
