+++ packages/perspective-viewer-d3fc/test/integration/barWidth.spec.js
