+++ packages/perspective-viewer-d3fc/test/integration/candlestick.spec.js
