+++ packages/perspective-viewer-d3fc/test/integration/scatter.spec.js
